+++ models/cards.js
@@ -3259,7 +3259,8 @@ if (Meteor.isServer) {
           boardId: paramBoardId,
           swimlaneId: paramSwimlaneId,
           archived: false,
-        }).map(function(doc) {
+        },
+        { sort: ['sort'] }).map(function(doc) {
           return {
             _id: doc._id,
             title: doc.title,
@@ -3270,6 +3271,7 @@ if (Meteor.isServer) {
             dueAt: doc.dueAt,
             endAt: doc.endAt,
             assignees: doc.assignees,
+            sort: doc.sort,
           };
         }),
       });
@@ -3301,16 +3303,19 @@ if (Meteor.isServer) {
         boardId: paramBoardId,
         listId: paramListId,
         archived: false,
-      }).map(function(doc) {
+      },
+      { sort: ['sort'] }).map(function(doc) {
         return {
           _id: doc._id,
           title: doc.title,
           description: doc.description,
+          swimlaneId: doc.swimlaneId,
           receivedAt: doc.receivedAt,
           startAt: doc.startAt,
           dueAt: doc.dueAt,
           endAt: doc.endAt,
           assignees: doc.assignees,
+          sort: doc.sort,
         };
       }),
     });
@@ -3555,6 +3560,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
       const newBoardId = req.body.newBoardId;
       const newSwimlaneId = req.body.newSwimlaneId;
       const newListId = req.body.newListId;
+      let updated = false;
       Authentication.checkBoardAccess(req.userId, paramBoardId);
 
       if (req.body.title) {
@@ -3572,6 +3578,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
             },
           },
         );
+        updated = true;
       }
       if (req.body.sort) {
         const newSort = req.body.sort;
@@ -3588,6 +3595,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
             },
           },
         );
+        updated = true;
       }
       if (req.body.parentId) {
         const newParentId = req.body.parentId;
@@ -3604,6 +3612,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
             },
           },
         );
+        updated = true;
       }
       if (req.body.description) {
         const newDescription = req.body.description;
@@ -3620,6 +3629,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
             },
           },
         );
+        updated = true;
       }
       if (req.body.color) {
         const newColor = req.body.color;
@@ -3632,6 +3642,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
           },
           { $set: { color: newColor } },
         );
+        updated = true;
       }
       if (req.body.vote) {
         const newVote = req.body.vote;
@@ -3650,6 +3661,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
           },
           { $set: { vote: newVote } },
         );
+        updated = true;
       }
       if (req.body.poker) {
         const newPoker = req.body.poker;
@@ -3675,6 +3687,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
           },
           { $set: { poker: newPoker } },
         );
+        updated = true;
       }
       if (req.body.labelIds) {
         let newlabelIds = req.body.labelIds;
@@ -3698,6 +3711,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
             },
           },
         );
+        updated = true;
       }
       if (req.body.requestedBy) {
         const newrequestedBy = req.body.requestedBy;
@@ -3710,6 +3724,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
           },
           { $set: { requestedBy: newrequestedBy } },
         );
+        updated = true;
       }
       if (req.body.assignedBy) {
         const newassignedBy = req.body.assignedBy;
@@ -3722,6 +3737,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
           },
           { $set: { assignedBy: newassignedBy } },
         );
+        updated = true;
       }
       if (req.body.receivedAt) {
         const newreceivedAt = req.body.receivedAt;
@@ -3734,6 +3750,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
           },
           { $set: { receivedAt: newreceivedAt } },
         );
+        updated = true;
       }
       if (req.body.startAt) {
         const newstartAt = req.body.startAt;
@@ -3746,6 +3763,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
           },
           { $set: { startAt: newstartAt } },
         );
+        updated = true;
       }
       if (req.body.dueAt) {
         const newdueAt = req.body.dueAt;
@@ -3758,6 +3776,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
           },
           { $set: { dueAt: newdueAt } },
         );
+        updated = true;
       }
       if (req.body.endAt) {
         const newendAt = req.body.endAt;
@@ -3770,6 +3789,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
           },
           { $set: { endAt: newendAt } },
         );
+        updated = true;
       }
       if (req.body.spentTime) {
         const newspentTime = req.body.spentTime;
@@ -3782,6 +3802,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
           },
           { $set: { spentTime: newspentTime } },
         );
+        updated = true;
       }
       if (req.body.isOverTime) {
         const newisOverTime = req.body.isOverTime;
@@ -3794,6 +3815,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
           },
           { $set: { isOverTime: newisOverTime } },
         );
+        updated = true;
       }
       if (req.body.customFields) {
         const newcustomFields = req.body.customFields;
@@ -3806,6 +3828,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
           },
           { $set: { customFields: newcustomFields } },
         );
+        updated = true;
       }
       if (req.body.members) {
         let newmembers = req.body.members;
@@ -3825,6 +3848,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
           },
           { $set: { members: newmembers } },
         );
+        updated = true;
       }
       if (req.body.assignees) {
         let newassignees = req.body.assignees;
@@ -3844,6 +3868,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
           },
           { $set: { assignees: newassignees } },
         );
+        updated = true;
       }
       if (req.body.swimlaneId) {
         const newParamSwimlaneId = req.body.swimlaneId;
@@ -3856,6 +3881,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
           },
           { $set: { swimlaneId: newParamSwimlaneId } },
         );
+        updated = true;
       }
       if (req.body.listId) {
         const newParamListId = req.body.listId;
@@ -3872,6 +3898,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
             },
           },
         );
+        updated = true;
 
         const card = ReactiveCache.getCard(paramCardId);
         cardMove(
@@ -3900,6 +3927,7 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
             },
           },
         );
+        updated = true;
 
         const card = ReactiveCache.getCard(paramCardId);
         cardMove(
@@ -3929,7 +3957,18 @@ JsonRoutes.add('GET', '/api/boards/:boardId/cards_count', function(
             },
           },
         );
+        updated = true;
       }
+    // Check if update is true or false
+    if (!updated) {
+      JsonRoutes.sendResult(res, {
+        code: 404,
+        data: {
+          message: 'Error',
+        },
+      });
+      return;
+    }
       JsonRoutes.sendResult(res, {
         code: 200,
         data: {